# promutil: rewrite sanitisation funcs for memory optimisation