--- a/pkg/promutil/migrate_test.go
+++ b/pkg/promutil/migrate_test.go
@@ -792,6 +792,159 @@ func TestBuildMetrics(t *testing.T) {
 	}
 }
 
+func Benchmark_BuildMetrics(b *testing.B) {
+	ts := time.Date(2024, time.January, 1, 0, 0, 0, 0, time.UTC)
+
+	data := []model.CloudwatchMetricResult{{
+		Context: &model.ScrapeContext{
+			Region:     "us-east-1",
+			AccountID:  "123456789012",
+			CustomTags: nil,
+		},
+		Data: []*model.CloudwatchData{
+			{
+				MetricName: "CPUUtilization",
+				MetricMigrationParams: model.MetricMigrationParams{
+					NilToZero:              true,
+					AddCloudwatchTimestamp: false,
+				},
+				Namespace: "AWS/ElastiCache",
+				GetMetricDataResult: &model.GetMetricDataResult{
+					Statistic: "Average",
+					Datapoint: aws.Float64(1),
+					Timestamp: ts,
+				},
+				Dimensions: []model.Dimension{
+					{
+						Name:  "CacheClusterId",
+						Value: "redis-cluster",
+					},
+				},
+				ResourceName: "arn:aws:elasticache:us-east-1:123456789012:cluster:redis-cluster",
+				Tags: []model.Tag{{
+					Key:   "managed_by",
+					Value: "terraform",
+				}},
+			},
+			{
+				MetricName: "FreeableMemory",
+				MetricMigrationParams: model.MetricMigrationParams{
+					NilToZero:              false,
+					AddCloudwatchTimestamp: false,
+				},
+				Namespace: "AWS/ElastiCache",
+				Dimensions: []model.Dimension{
+					{
+						Name:  "CacheClusterId",
+						Value: "redis-cluster",
+					},
+				},
+				GetMetricDataResult: &model.GetMetricDataResult{
+					Statistic: "Average",
+					Datapoint: aws.Float64(2),
+					Timestamp: ts,
+				},
+				ResourceName: "arn:aws:elasticache:us-east-1:123456789012:cluster:redis-cluster",
+				Tags: []model.Tag{{
+					Key:   "managed_by",
+					Value: "terraform",
+				}},
+			},
+			{
+				MetricName: "NetworkBytesIn",
+				MetricMigrationParams: model.MetricMigrationParams{
+					NilToZero:              true,
+					AddCloudwatchTimestamp: false,
+				},
+				Namespace: "AWS/ElastiCache",
+				Dimensions: []model.Dimension{
+					{
+						Name:  "CacheClusterId",
+						Value: "redis-cluster",
+					},
+				},
+				GetMetricDataResult: &model.GetMetricDataResult{
+					Statistic: "Average",
+					Datapoint: aws.Float64(3),
+					Timestamp: ts,
+				},
+				ResourceName: "arn:aws:elasticache:us-east-1:123456789012:cluster:redis-cluster",
+				Tags: []model.Tag{{
+					Key:   "managed_by",
+					Value: "terraform",
+				}},
+			},
+			{
+				MetricName: "NetworkBytesOut",
+				MetricMigrationParams: model.MetricMigrationParams{
+					NilToZero:              true,
+					AddCloudwatchTimestamp: true,
+				},
+				Namespace: "AWS/ElastiCache",
+				Dimensions: []model.Dimension{
+					{
+						Name:  "CacheClusterId",
+						Value: "redis-cluster",
+					},
+				},
+				GetMetricDataResult: &model.GetMetricDataResult{
+					Statistic: "Average",
+					Datapoint: aws.Float64(4),
+					Timestamp: ts,
+				},
+				ResourceName: "arn:aws:elasticache:us-east-1:123456789012:cluster:redis-cluster",
+				Tags: []model.Tag{{
+					Key:   "managed_by",
+					Value: "terraform",
+				}},
+			},
+		},
+	}}
+
+	var labels map[string]model.LabelSet
+	var err error
+
+	b.ReportAllocs()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_, labels, err = BuildMetrics(data, false, logging.NewNopLogger())
+	}
+
+	expectedLabels := map[string]model.LabelSet{
+		"aws_elasticache_cpuutilization_average": {
+			"account_id":               {},
+			"name":                     {},
+			"region":                   {},
+			"dimension_CacheClusterId": {},
+			"tag_managed_by":           {},
+		},
+		"aws_elasticache_freeable_memory_average": {
+			"account_id":               {},
+			"name":                     {},
+			"region":                   {},
+			"dimension_CacheClusterId": {},
+			"tag_managed_by":           {},
+		},
+		"aws_elasticache_network_bytes_in_average": {
+			"account_id":               {},
+			"name":                     {},
+			"region":                   {},
+			"dimension_CacheClusterId": {},
+			"tag_managed_by":           {},
+		},
+		"aws_elasticache_network_bytes_out_average": {
+			"account_id":               {},
+			"name":                     {},
+			"region":                   {},
+			"dimension_CacheClusterId": {},
+			"tag_managed_by":           {},
+		},
+	}
+
+	require.NoError(b, err)
+	require.Equal(b, expectedLabels, labels)
+}
+
 // replaceNaNValues replaces any NaN floating-point values with a marker value (54321.0)
 // so that require.Equal() can compare them. By default, require.Equal() will fail if any
 // struct values are NaN because NaN != NaN