Rewrite `sanitize` to attempt to modify the string in place, if possible.
Rewrite `splitString` to avoid using a regex.

```
% benchstat before.txt after.txt
goos: darwin
goarch: arm64
pkg: github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/promutil
                 │ before.txt  │              after.txt              │
                 │   sec/op    │   sec/op     vs base                │
_BuildMetrics-12   201.1µ ± 2%   114.5µ ± 6%  -43.07% (p=0.000 n=10)

                 │  before.txt  │              after.txt               │
                 │     B/op     │     B/op      vs base                │
_BuildMetrics-12   5.750Ki ± 1%   4.281Ki ± 0%  -25.55% (p=0.000 n=10)

                 │ before.txt  │             after.txt              │
                 │  allocs/op  │ allocs/op   vs base                │
_BuildMetrics-12   168.00 ± 0%   95.00 ± 0%  -43.45% (p=0.000 n=10)
```